// src/types/BackgroundOperationPhase.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BackgroundOperationPhase = "scanningPackages" | "mergingLoadOrder" | "savingLoadOrder" | "parsingCsv" | "applyingLoadOrder" | "loadingLoadOrder" | "resolvingLoadOrder" | "updatingServerMetadata" | "hashingPackage" | "hashingFiles" | "writingManifest" | "writingLaunchCfg" | "downloadingRuntime" | "copyingRuntime" | "extractingRuntime" | "inspectingRuntime" | "complete";